+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Not released][Not released]
 
+### Fixed
+- `getClientRect` calculations
+
 ## [1.2.0][2016-09-15]
 
 ## Added